+++ sycl/pi_win_proxy_loader/pi_win_proxy_loader.cpp
@@ -39,83 +39,8 @@
 #include <string>
 
 #include "pi_win_proxy_loader.hpp"
-
+#include <sycl/detail/windows_os_utils.hpp>
 #ifdef _WIN32
-
-// ------------------------------------
-
-static constexpr const char *DirSep = "\\";
-
-// cribbed from sycl/source/detail/os_util.cpp
-std::string getDirName(const char *Path) {
-  std::string Tmp(Path);
-  // Remove trailing directory separators
-  Tmp.erase(Tmp.find_last_not_of("/\\") + 1, std::string::npos);
-
-  size_t pos = Tmp.find_last_of("/\\");
-  if (pos != std::string::npos)
-    return Tmp.substr(0, pos);
-
-  // If no directory separator is present return initial path like dirname does
-  return Tmp;
-}
-
-// cribbed from sycl/source/detail/os_util.cpp
-// TODO: Just inline it.
-using OSModuleHandle = intptr_t;
-static constexpr OSModuleHandle ExeModuleHandle = -1;
-static OSModuleHandle getOSModuleHandle(const void *VirtAddr) {
-  HMODULE PhModule;
-  DWORD Flag = GET_MODULE_HANDLE_EX_FLAG_FROM_ADDRESS |
-               GET_MODULE_HANDLE_EX_FLAG_UNCHANGED_REFCOUNT;
-  auto LpModuleAddr = reinterpret_cast<LPCSTR>(VirtAddr);
-  if (!GetModuleHandleExA(Flag, LpModuleAddr, &PhModule)) {
-    // Expect the caller to check for zero and take
-    // necessary action
-    return 0;
-  }
-  if (PhModule == GetModuleHandleA(nullptr))
-    return ExeModuleHandle;
-  return reinterpret_cast<OSModuleHandle>(PhModule);
-}
-
-// cribbed from sycl/source/detail/os_util.cpp
-/// Returns an absolute path where the object was found.
-std::string getCurrentDSODir() {
-  char Path[MAX_PATH];
-  Path[0] = '\0';
-  Path[sizeof(Path) - 1] = '\0';
-  auto Handle = getOSModuleHandle(reinterpret_cast<void *>(&getCurrentDSODir));
-  DWORD Ret = GetModuleFileNameA(
-      reinterpret_cast<HMODULE>(ExeModuleHandle == Handle ? 0 : Handle),
-      reinterpret_cast<LPSTR>(&Path), sizeof(Path));
-  assert(Ret < sizeof(Path) && "Path is longer than PATH_MAX?");
-  assert(Ret > 0 && "GetModuleFileNameA failed");
-  (void)Ret;
-
-  BOOL RetCode = PathRemoveFileSpecA(reinterpret_cast<LPSTR>(&Path));
-  assert(RetCode && "PathRemoveFileSpecA failed");
-  (void)RetCode;
-
-  return Path;
-}
-
-std::filesystem::path getCurrentDSODirPath() {
-  wchar_t Path[MAX_PATH];
-  auto Handle = getOSModuleHandle(reinterpret_cast<void *>(&getCurrentDSODir));
-  DWORD Ret = GetModuleFileName(
-      reinterpret_cast<HMODULE>(ExeModuleHandle == Handle ? 0 : Handle),
-      reinterpret_cast<LPWSTR>(&Path), sizeof(Path));
-  assert(Ret < sizeof(Path) && "Path is longer than PATH_MAX?");
-  assert(Ret > 0 && "GetModuleFileName failed");
-  (void)Ret;
-
-  BOOL RetCode = PathRemoveFileSpec(reinterpret_cast<LPWSTR>(&Path));
-  assert(RetCode && "PathRemoveFileSpec failed");
-  (void)RetCode;
-
-  return std::filesystem::path(std::wstring(Path));
-}
 
 // these are cribbed from include/sycl/detail/pi.hpp
 // a new plugin must be added to both places.
@@ -164,7 +89,7 @@ void preloadLibraries() {
   }
 
   // this path duplicates sycl/detail/pi.cpp:initializePlugins
-  std::filesystem::path LibSYCLDir = getCurrentDSODirPath();
+  std::filesystem::path LibSYCLDir = sycl::detail::getCurrentDSODirPath();
 
   MapT &dllMap = getDllMap();
